requirements.txt
accelerate==0.22.0
appdirs==1.4.4
asttokens==2.4.1
bitsandbytes==0.41.1
black==23.11.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.12.3
fonttools==4.44.0
fsspec==2023.9.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.3.1
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
peft==0.5.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scipy==1.11.3
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
sklearn==0.0.post7
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
termcolor==2.3.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.33.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Wand==0.6.11
wandb==0.15.12
wcwidth==0.2.9
widgetsnbextension==4.0.9
